+++ src/main/java/io/ezra/lending/repos/LoanApplicationRepo.java
@@ -3,5 +3,7 @@ package io.ezra.lending.repos;
 import io.ezra.lending.entities.LoanApplication;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LoanApplicationRepo extends JpaRepository<LoanApplication, Integer> {
+import java.math.BigInteger;
+
+public interface LoanApplicationRepo extends JpaRepository<LoanApplication, BigInteger> {
 }
